% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_analyses.R
\name{coef_change_detection}
\alias{coef_change_detection}
\title{Detect coefficient change for target genes after adjusting for estimated cell proportions
The following code are used to detect coefficient change for target genes after adjusting for estimated cell proportions}
\usage{
coef_change_detection(X_assoc, Y, PCs_active, Cell_prop)
}
\arguments{
\item{X_assoc}{tissue-specific gene expression profile for all target genes}

\item{Y}{the target trait}

\item{PCs_active}{top PCs used to adjust for population stratificataion}

\item{Cell_prop}{the estimated cell proportions for the target tissue}
}
\value{
a result matrix with 15 columns, i.e., Genes, Estimates, Pvalues, Pvalues_BH, Pvalues_Bonferroni etc.
}
\description{
Detect coefficient change for target genes after adjusting for estimated cell proportions
The following code are used to detect coefficient change for target genes after adjusting for estimated cell proportions
}
