% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_analyses.R
\name{cTWAS}
\alias{cTWAS}
\title{Get cell type specific trait-associated genes for the target traits}
\usage{
cTWAS(X_assoc, Y, PCs_active)
}
\arguments{
\item{X_assoc}{cell type specific gene expression profiles for all genes}

\item{Y}{the target trait}

\item{PCs_active}{top PCs used to adjust for population stratificataion}
}
\value{
a result matrix with 5 columns, i.e., Genes, Estimates, Pvalues, Pvalues_BH, Pvalues_Bonferroni
}
\description{
The following code are used to identify cell type specific associated genes for the target trait
}
