% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_estimation.R
\name{.onLoad}
\alias{.onLoad}
\title{Load the scaden_py module for deconvolution}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{sc_dat}{the reference single cell data.}

\item{bulk_dat}{the bulk data to be deconvoluted.}
}
\value{
a result matrix with 3 columns, i.e., Genes, zMin, Assoc
}
\description{
The .onLoad code is used to load the scaden_py module for deconvolution,
}
