% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCM.R
\name{gcm_test}
\alias{gcm_test}
\title{Revise the original gcm.test from R package "GeneralisedCovarianceMeasure"}
\usage{
gcm_test(
  X = NULL,
  Y = NULL,
  Z = NULL,
  alpha = 0.05,
  regr.method = "xgboost",
  regr.pars = list(),
  plot.residuals = FALSE,
  nsim = 499L,
  resid.XonZ = NULL,
  resid.YonZ = NULL
)
}
\arguments{
\item{X}{A(nxp)-dimensional matrix (or data frame) with n observations of p variables.}

\item{Y}{A(nxp)-dimensional matrix (or data frame) with n observations of p variables.}

\item{Z}{A(nxp)-dimensional matrix (or data frame) with n observations of p variables.}

\item{alpha}{Significance level of the test.}

\item{regr.method}{A string indicating the regression method that is used. Currently implemented are "gam", "xgboost", "kernel.ridge". The regression is performed only if not both resid.XonZ and resid.YonZ are set to NULL.}

\item{regr.pars}{Some regression methods require a list of additional options}

\item{plot.residuals}{A Boolean indicating whether some plots should be shown.}

\item{nsim}{An integer indicating the number of bootstrap samples used to approximate the null distribution of the test statistic.}

\item{resid.XonZ}{It is possible to directly provide the residuals instead of performing a regression. If set to NULL, the regression method specified in regr.method is used.}

\item{resid.YonZ}{It is possible to directly provide the residuals instead of performing a regression. If set to NULL, the regression method specified in regr.method is used.}
}
\value{
a list containing P-value, R, test.statistic and whether reject
}
\description{
This code is to revise the original gcm.test from R package "GeneralisedCovarianceMeasure"
}
