% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_analyses.R
\name{get_causal_genes}
\alias{get_causal_genes}
\title{Get the causal genes using the proposed PC-simple algorithm}
\usage{
get_causal_genes(data, outcome, alpha, num_workers)
}
\arguments{
\item{data}{unconfounded cell type specific gene expression profiles for all genes}

\item{outcome}{the target trait}

\item{alpha}{the cutoff used to define causally relevant genes, we set alpha to 0.05 in this study, but you can adaptively adjust this cutoff}

\item{num_workers}{the number of cores used to calculate causally relevant genes, we set it to 10 in our study. Again you can adjust this to adapt to your device}
}
\value{
a result matrix with 3 columns, i.e., Genes, zMin, Assoc
}
\description{
The following code are used to identify cell type specific causal genes for the target trait
}
