% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enigma_infer_CTS.R
\name{infer_CTS_profile}
\alias{infer_CTS_profile}
\title{Infer the CTS profile using ENIGMA}
\usage{
infer_CTS_profile(bulk_dat, sc_dat, prop, ...)
}
\arguments{
\item{bulk_dat}{bulk RNA-seq dataset, we expect the bulk data index should match the propotion data}

\item{sc_dat}{single cell RNA-seq dataset.}

\item{prop}{estimated proportion of the bulk RNA-seq dataset.}
}
\value{
a list containing the estimated CTS profile
}
\description{
The following code is to use the ENIGMA to infer the CTS profile for different cell types.
}
