% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-GCM.R
\name{InvariantGCM}
\alias{InvariantGCM}
\title{I-GCM is designed to accurately identify direct causal variable set for the target variable}
\usage{
InvariantGCM(
  X_assoc,
  Y,
  envir,
  change_alpha = 0.05,
  change_gcm = 0,
  num_workers,
  verbose = FALSE
)
}
\arguments{
\item{X_assoc}{all input variables excluding the environment and target variable, with each row represents one subject}

\item{Y}{target variable vector for the same set of subjects}

\item{envir}{environment variable vector for the same set of subjects}

\item{change_alpha}{p-value cutoff used to detect significant GCM change, default alpha=0.05}

\item{change_gcm}{absolute gcm change threshold for significant change, default change_gcm=0}

\item{num_workers}{number of threads used for feature selection}

\item{verbose}{if verbose=TRUE, the details of the analysis will be print out. The default value is false}
}
\value{
res a list stored all related results, which includes candidate_features, resid_mat and all_candidate_causal_set; candidate_features: the preselected candidate causal variables from PC-simple ranked by zscore in descending order; resid_mat: the residual matrix with each column stores the residuals for all subjects in each performed test; all_candidate_causal_set: number of causal variables(The top ranked variables from candidate_features) for each identified
}
\description{
The invariantGCM function is to accurately identify direct causal variable set for the target variable
}
