% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cellmarkers.R
\name{merge_and_rank_genes}
\alias{merge_and_rank_genes}
\title{Re-rank genes based on zMin}
\usage{
merge_and_rank_genes(df1, df2)
}
\arguments{
\item{df1}{dataframe with three columns: genes, assoc and zMin to store the identifed caual genes from environment 1}

\item{df2}{dataframe with three columns: genes, assoc and zMin to store the identifed caual genes from environment 2}
}
\value{
final_df re-rank the merged df in descending order(zMin), with zMin indicates variable importance
}
\description{
merge_and_rank_genes is designed to re-rank genes based on zMin
}
