% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCSelect_Parallel_Update.R
\name{PCSelect_Parallel}
\alias{PCSelect_Parallel}
\title{Identify the causal gene}
\usage{
PCSelect_Parallel(
  y,
  dm,
  method = c("parallel"),
  mem.efficient = FALSE,
  num_workers,
  alpha,
  corMat = NA,
  corMethod = "standard",
  max_ord = 5,
  verbose = FALSE,
  directed = FALSE
)
}
\arguments{
\item{y}{Response Vector (length(y)=nrow(dm))}

\item{dm}{Data matrix (rows: samples, cols: nodes)}

\item{method}{Character string specifying method; the default, "parallel" provides an parallelised method to implement all the conditional independence tests.}

\item{mem.efficient}{If TRUE, uses less amount of memory at any time point while running the algorithm}

\item{num_workers}{The numbers of cores CPU to run the algorithm}

\item{alpha}{Significance level of individual partial correlation tests}

\item{corMat}{Correlation Matrix}

\item{corMethod}{"standard" or "Qn" for standard or robust correlation estimation}

\item{verbose}{0-no output, 1-small output, 2-details}

\item{directed}{Logical; should the output graph be directed?}
}
\value{
A list object containing all target cell types and their corresponding cell markers.
}
\description{
The following code is used to find columns in dm, that have nonzero parcor with y given any other set of columns in dm
}
