% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCM.R
\name{stat_change_detection}
\alias{stat_change_detection}
\title{Calculate the covariance between the calculated GCMs from two different GCM tests}
\usage{
stat_change_detection(R1, R2, N)
}
\arguments{
\item{R1}{derived residual product for the 1st test}

\item{R2}{derived residual product for the 2nd test}

\item{N}{sample size}
}
\value{
calculated covariance
}
\description{
The following code is used to calculate the covariance between the calculated GCMs from two different GCM tests.
}
