% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cellmarkers.R
\name{get_cellmarkers}
\alias{get_cellmarkers}
\title{Get the cell markers for cell deconvolution using ICP approach}
\usage{
get_cellmarkers(label, data)
}
\arguments{
\item{label}{the corresponding cell type labels for the input data}

\item{data}{input normalized scRNA-seq data(count data would also be OK), numeric data type}
}
\value{
A list object containing all target cell types and their corresponding cell markers.
}
\description{
The following code are used to identify environment specific cellmarkers.
}
