% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_estimation.R
\name{get_proportion}
\alias{get_proportion}
\title{Obtain the proportion of the cell types of a target}
\usage{
get_proportion(sc_dat, bulk_dat, marker_genes, ...)
}
\arguments{
\item{sc_dat}{the reference single cell data.}

\item{bulk_dat}{the bulk data to be deconvoluted.}

\item{marker_genes}{marker genes for deconvolution.}

\item{...}{arguments from the scaden deconvolution.}
}
\value{
a result matrix with 3 columns, i.e., Genes, zMin, Assoc
}
\description{
The following code are used to identify cell type specific causal genes for the target trait
}
